refactor: fix filename typos, type checking, rename files
