--- src/components/Headers/index.ts
@@ -1,3 +1,0 @@
-import Headers from './Headers';
-
-export default Headers;
