+++ src/components/HeadersSection/HeadersSection.module.scss
