+++ src/components/HeadersSection/index.ts
@@ -1,0 +1,3 @@
+import Headers from './HeadersSection';
+
+export default Headers;
