+++ src/components/MainEndpointInput/index.ts
@@ -1,0 +1,3 @@
+import MainEndpointInput from './MainEndpointInput';
+
+export default MainEndpointInput;
