--- src/components/Variables/index.ts
@@ -1,3 +1,0 @@
-import Variables from './Variables';
-
-export default Variables;
