+++ src/components/VariablesSection/VariablesSection.module.scss
