+++ src/components/VariablesSection/VariablesSection.tsx
@@ -1,10 +1,12 @@
 import { useActions, useAppSelector } from 'hooks/redux-hooks';
 
+import { useLocalization } from 'shared/context/LocalizationContext';
 import { RootState } from 'shared/store/store';
 
-import styles from './Variables.module.scss';
+import styles from './VariablesSection.module.scss';
 
 function Variables() {
+  const { t } = useLocalization();
   const selectVariables = (state: RootState) => state.editor.variables;
   const variables = useAppSelector(selectVariables);
   const { setVariables } = useActions();
@@ -16,7 +18,7 @@ function Variables() {
   return (
     <textarea
       className={styles.variables__textarea}
-      placeholder="Enter variables"
+      placeholder={t('placeholders.variables')}
       value={variables}
       onChange={handleInputChange}
     />
