+++ src/components/VariablesSection/index.ts
@@ -1,0 +1,3 @@
+import Variables from './VariablesSection';
+
+export default Variables;
